Fix TC kernel crash: Y_smem must be __shared__ for WMMA on Turing
BUG 1 (crash): Y_float_smem and Y_smem were local arrays. WMMA's
store_matrix_sync on sm_75 (Turing) requires shared memory — the
hardware scatter path cannot write to per-thread local memory.
Added __shared__ and __syncthreads() after store.

BUG 2 (docs): Docstring still said int64, but pack_tensor returns int32
since the earlier stride fix.

BUG 3 (minor): K-padding works accidentally due to bounds check.

+++ kernels/packed_ternary/pack_forward.py
@@ -4,7 +4,7 @@ pack_forward.py — Packed ternary × FP16 forward GEMM via CUDA.
 Provides:
     packed_ternary_forward(W_packed, X_fp16) → Y_fp16
 
-The weight tensor W_packed is a (out_features, stride_words) int64 tensor
+The weight tensor W_packed is a (out_features, stride_words) int32 tensor
 packed with 16 ternary values per word (see pack_tensor).
 
 Reference (for testing):
